+++ src/controller-generator/controller-generator.controller.ts
@@ -1,3 +1,4 @@
+import { FileGenerated } from '../shared/constants';
 import { Controller, Post, Body } from '@nestjs/common';
 import { ControllerGeneratorService } from './controller-generator.service';
 import { CreateControllerDto } from '../shared/models/create-controller.dto';
@@ -7,7 +8,10 @@ export class ControllerGeneratorController {
   constructor(private readonly controllerGeneratorService: ControllerGeneratorService) {}
 
   @Post()
-  public createController(@Body() createControllerDto: CreateControllerDto) {
-    return this.controllerGeneratorService.generateController(createControllerDto);
+  public createController(@Body() createControllerDto: CreateControllerDto): FileGenerated {
+    const code = this.controllerGeneratorService.generateController(createControllerDto);
+    const path = this.controllerGeneratorService.getFilePath(createControllerDto.entityName);
+
+    return { code, path };
   }
 }