+++ src/module-generator/module-generator.service.ts
@@ -1,10 +1,16 @@
+import { getFilePath } from '../templates';
 import { Injectable } from '@nestjs/common';
+import { FileType } from '../shared/constants';
+import { generateClass } from '../templates/module.template';
 import { CreateModuleDto } from '../shared/models/create-module.dto';
-import { generateClass, discoverModuleImports } from '../templates/module.template';
 
 @Injectable()
 export class ModuleGeneratorService {
   generateModule(createModuleDto: CreateModuleDto) {
     return generateClass(createModuleDto);
+  }
+
+  getFilePath(entityName: string) {
+    return getFilePath(entityName, FileType.module);
   }
 }