+++ src/service-generator/service-generator.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@nestjs/common';
-import { Layer } from '../shared/constants';
-import { MountLayerParams, mountLayerFile, getDefaultLayerBellow } from '../templates';
+import { Layer, FileType } from '../shared/constants';
 import { CreateServiceDto } from '../shared/models/create-service.dto';
 import { discoverServiceImports } from '../templates/service.template';
+import { MountLayerParams, mountLayerFile, getDefaultLayerBellow, getFilePath } from '../templates';
 
 @Injectable()
 export class ServiceGeneratorService {
@@ -24,5 +24,9 @@ export class ServiceGeneratorService {
 
     const fileContent = mountLayerFile(params);
     return fileContent;
+  }
+
+  getFilePath(entityName: string) {
+    return getFilePath(entityName, FileType.service);
   }
 }